requirement.txt
Django==4.0
djangorestframework>=3.13.1
markdown>=3.3.7
django-filter>=22.1
psycopg2-binary>=2.8
django-environ>=0.8.1
django-cors-headers>=3.13.0
pytz>=2022.1
six>=1.16.0
djangorestframework-simplejwt>=5.2.0
PyJWT==v1.7.1

